Reading and sending depth instead of pressure

+++ include/Communication.h
@@ -8,5 +8,5 @@ extern StaticJsonDocument<48> outDoc;
 // TESTING VARIABLE
 extern char json[];
 
-void sendToRaspberry(float arg1, int arg2, bool arg3);
+void sendToRaspberry(float arg1, float arg2, bool arg3);
 void receiveFromRaspberry();
